# lib/api-zod/src/generated/types/grammarRule.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kazakh Language LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { GrammarExample } from './grammarExample';

export interface GrammarRule {
  id: number;
  title: string;
  level: string;
  topic: string;
  description: string;
  examples: GrammarExample[];
  /** @nullable */
  relatedRules?: number[] | null;
}