# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kazakh Language LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonInputType } from './lessonInputType';

export interface LessonInput {
  courseId: number;
  title: string;
  type: LessonInputType;
  order: number;
  duration: number;
  xpReward: number;
  content: string;
}