# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kazakh Language LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  name?: string;
  bio?: string;
  educationLevel?: string;
  avatar?: string;
}