# lib/api-zod/src/generated/types/vocabWordInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kazakh Language LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface VocabWordInput {
  kazakh: string;
  russian: string;
  english?: string;
  category: string;
  level: string;
  examples?: string[];
}